Move entities to `entities` folder
