+++ src/auth/auth.service.ts
@@ -4,7 +4,7 @@ import { InjectRepository } from "@nestjs/typeorm";
 import * as bcrypt from "bcrypt";
 import { TokenExpiredError } from "jsonwebtoken";
 import { Repository } from "typeorm";
-import { User } from "../users/user.entity";
+import { User } from "../users/entities/user.entity";
 import { UsersService } from "../users/users.service";
 import { RefreshToken } from "./refresh-token.entity";
