+++ src/auth/current-user.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from "@nestjs/common";
-import { User } from "../users/user.entity";
+import { User } from "../users/entities/user.entity";
 
 export const CurrentUser = createParamDecorator<
   unknown,
