+++ src/auth/dto/register-user.payload.ts
@@ -1,6 +1,6 @@
 import { Field, ObjectType } from "@nestjs/graphql";
 import { UserObject } from "../../users/dto/user.object";
-import { User } from "../../users/user.entity";
+import { User } from "../../users/entities/user.entity";
 
 @ObjectType()
 export class RegisterUserPayload {
