+++ src/auth/gql-current-user.decorator.ts
@@ -1,6 +1,6 @@
 import { createParamDecorator, ExecutionContext } from "@nestjs/common";
 import { GqlExecutionContext } from "@nestjs/graphql";
-import { User } from "../users/user.entity";
+import { User } from "../users/entities/user.entity";
 
 export const GqlCurrentUser = createParamDecorator<
   unknown,
