+++ src/auth/refresh-token.entity.ts
@@ -1,4 +1,4 @@
-import { User } from "../users/user.entity";
+import { User } from "../users/entities/user.entity";
 import {
   Column,
   CreateDateColumn,
