+++ src/database/seeders/flairs/flairs.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { Flair } from "src/flairs/flair.entity";
+import { Flair } from "src/flairs/entities/flair.entity";
 import { FlairsSeederService } from "./flairs.service";
 
 @Module({
