+++ src/flairs/flairs.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { Flair } from "./flair.entity";
+import { Flair } from "./entities/flair.entity";
 import { FlairsService } from "./flairs.service";
 import { FlairsResolver } from './flairs.resolver';
