+++ src/posts/posts.module.ts
@@ -2,7 +2,7 @@ import { forwardRef, Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { FlairsModule } from "src/flairs/flairs.module";
 import { UsersModule } from "src/users/users.module";
-import { Post } from "./post.entity";
+import { Post } from "./entities/post.entity";
 import { PostsResolver } from "./posts.resolver";
 import { PostsService } from "./posts.service";
