+++ src/users/dto/user.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from "@nestjs/swagger";
 import { Exclude, Expose } from "class-transformer";
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { User } from "../user.entity";
+import { User } from "../entities/user.entity";
 
 @Exclude()
 export class UserDto {
