+++ src/users/users.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
-import { User } from "./user.entity";
+import { User } from "./entities/user.entity";
 
 interface FindOneArgs {
   id?: number;
